Modified Ordered Random Forest

Nonparametric estimator of the ordered choice model using random forests. The estimator modifies a standard random forest splitting criterion to build a collection of forests, each estimating the conditional probability of a single class.

morf(
  y = NULL,
  X = NULL,
  honesty = FALSE,
  honesty.fraction = 0.5,
  inference = FALSE,
  alpha = 0,
  n.trees = 2000,
  mtry = ceiling(sqrt(ncol(X))),
  min.node.size = 5,
  max.depth = 0,
  replace = FALSE,
  sample.fraction = ifelse(replace, 1, 0.5)
)

Arguments

y

Outcome vector.

X

Covariate matrix (no intercept).

honesty

Whether to grow honest forests.

honesty.fraction

Fraction of honest sample. Ignored if honesty = FALSE.

inference

Whether to extract weights and compute standard errors. The weights extraction considerably slows down the routine. honesty = TRUE is required for valid inference.

alpha

Controls the balance of each split. Each split leaves at least a fraction alpha of observations in the parent node on each side of the split.

n.trees

Number of trees.

mtry

Number of covariates to possibly split at in each node. Default is the square root of the number of covariates.

min.node.size

Minimal node size.

max.depth

Maximal tree depth. A value of 0 corresponds to unlimited depth, 1 to "stumps" (one split per tree).

replace

If TRUE, grow trees on bootstrap subsamples. Otherwise, trees are grown on random subsamples drawn without replacement.

sample.fraction

Fraction of observations to sample.

Value

Object of class morf.

See also

Author

Riccardo Di Francesco

Examples

# \donttest{
## Load data from orf package.
set.seed(1986)

library(orf)
data(odata)

y <- as.numeric(odata[, 1])
X <- as.matrix(odata[, -1])

## Training-test split.
train_idx <- sample(seq_len(length(y)), length(y)/2)

y_tr <- y[train_idx]
X_tr <- X[train_idx, ]

y_test <- y[-train_idx]
X_test <- X[-train_idx, ]

## Fit morf on training sample.
forests <- morf(y_tr, X_tr)

## We have compatibility with generic S3-methods.
print(forests)
#> Call: 
#> morf(y_tr, X_tr) 
#> 
#> Data info: 
#> Full sample size:   500 
#> N. covariates:      5 
#> Classes:            1 2 3 
#> 
#> Relative variable importance: 
#>    X1    X2    X3    X4 
#> 0.410 0.190 0.104 0.295 
#> 
#> Tuning parameters: 
#> N. trees:           2000 
#> mtry:               2 
#> min.node.size       5 
#> Subsampling scheme: No replacement 
#> Honesty:            FALSE 
#> Honest fraction:    0
summary(forests)
#> Call: 
#> morf(y_tr, X_tr) 
#> 
#> Data info: 
#> Full sample size:   500 
#> N. covariates:      5 
#> Classes:            1 2 3 
#> 
#> Relative variable importance: 
#>    X1    X2    X3    X4 
#> 0.410 0.190 0.104 0.295 
#> 
#> Tuning parameters: 
#> N. trees:           2000 
#> mtry:               2 
#> min.node.size       5 
#> Subsampling scheme: No replacement 
#> Honesty:            FALSE 
#> Honest fraction:    0
predictions <- predict(forests, X_test)
head(predictions$probabilities)
#>          P(Y=1)    P(Y=2)     P(Y=3)
#> [1,] 0.85140961 0.1185089 0.03008149
#> [2,] 0.70315449 0.2661564 0.03068916
#> [3,] 0.36277830 0.2800366 0.35718506
#> [4,] 0.01424455 0.1056402 0.88011530
#> [5,] 0.11414762 0.4527127 0.43313970
#> [6,] 0.03066153 0.2633758 0.70596271
table(y_test, predictions$classification)
#>       
#> y_test   1   2   3
#>      1 104  41  20
#>      2  61  59  46
#>      3  38  45  86

## Compute standard errors. This requires honest forests.
honest_forests <- morf(y_tr, X_tr, honesty = TRUE, inference = TRUE)
head(honest_forests$predictions$standard.errors)# }
#>          P(Y=1)     P(Y=2)     P(Y=3)
#> [1,] 0.03741186 0.05767769 0.14603833
#> [2,] 0.12405204 0.05537547 0.06942325
#> [3,] 0.03946996 0.04222065 0.24597406
#> [4,] 0.12612915 0.13520565 0.09198743
#> [5,] 0.13156758 0.14102797 0.02118072
#> [6,] 0.12020052 0.08409542 0.02783215