Prediction Method for morf Objects

Prediction method for class morf.

# S3 method for morf
predict(object, data = NULL, type = "response", ...)

Arguments

object

An morf object.

data

Data set of class data.frame. It must contain at least the same covariates used to train the forests. If data is NULL, then object$full_data is used.

type

Type of prediction. Either "response" or "terminalNodes".

...

Further arguments passed to or from other methods.

Value

Desired predictions.

Details

If type == "response", the routine returns the predicted conditional class probabilities and the predicted class labels. If forests are honest, the predicted probabilities are honest.

If type == "terminalNodes", the IDs of the terminal node in each tree for each observation in data are returned.

See also

Author

Riccardo Di Francesco

Examples

# \donttest{
## Load data from orf package.
set.seed(1986)

library(orf)
data(odata)

y <- as.numeric(odata[, 1])
X <- as.matrix(odata[, -1])

## Training-test split.
train_idx <- sample(seq_len(length(y)), length(y)/2)

y_tr <- y[train_idx]
X_tr <- X[train_idx, ]

y_test <- y[-train_idx]
X_test <- X[-train_idx, ]

## Fit morf on training sample.
forests <- morf(y_tr, X_tr)

## Predict on test sample.
predictions <- predict(forests, X_test)
head(predictions$probabilities)
#>          P(Y=1)    P(Y=2)     P(Y=3)
#> [1,] 0.85140961 0.1185089 0.03008149
#> [2,] 0.70315449 0.2661564 0.03068916
#> [3,] 0.36277830 0.2800366 0.35718506
#> [4,] 0.01424455 0.1056402 0.88011530
#> [5,] 0.11414762 0.4527127 0.43313970
#> [6,] 0.03066153 0.2633758 0.70596271
predictions$classification
#>   [1] 1 1 1 3 2 3 3 3 2 3 3 2 1 1 3 3 1 3 2 3 1 1 3 3 2 1 2 2 3 3 2 1 3 3 3 3 1
#>  [38] 1 1 2 2 1 2 3 3 3 1 1 3 1 2 2 1 1 3 2 1 1 2 1 1 2 3 1 2 3 1 3 1 2 3 3 2 3
#>  [75] 1 1 1 1 3 3 1 1 3 2 1 3 3 3 1 1 3 1 1 1 1 2 1 1 3 2 2 3 3 2 1 1 2 1 2 1 1
#> [112] 1 3 2 3 2 1 2 3 3 2 2 2 1 1 2 2 2 3 1 1 3 2 3 1 3 2 3 3 3 1 3 3 1 2 1 2 2
#> [149] 1 2 1 1 2 3 2 1 2 1 3 2 1 3 2 1 3 1 2 2 1 1 2 2 3 1 2 1 2 2 1 1 1 3 1 3 1
#> [186] 3 3 1 1 3 1 3 1 2 1 3 1 1 2 1 1 1 1 2 2 3 3 1 3 3 3 1 2 3 2 1 1 2 3 1 3 2
#> [223] 1 2 2 1 1 2 2 1 3 2 3 2 2 3 1 2 3 2 2 1 2 3 1 1 3 3 2 2 3 2 1 3 2 1 2 2 2
#> [260] 1 2 2 1 1 3 1 3 1 1 1 3 3 1 3 2 2 2 3 1 1 2 2 2 3 1 3 2 1 1 2 1 3 3 3 1 1
#> [297] 1 2 1 1 3 1 1 3 3 1 1 2 1 3 1 2 1 3 1 2 3 1 1 3 1 1 3 3 2 1 1 3 3 2 1 1 2
#> [334] 1 3 3 3 3 1 1 1 1 3 1 2 1 3 3 2 1 1 1 2 1 3 2 1 1 2 3 2 2 2 2 3 3 2 2 1 1
#> [371] 1 3 1 3 1 3 2 2 1 1 1 1 3 3 2 1 3 1 2 3 1 2 2 1 2 2 2 1 3 3 1 2 3 2 2 1 2
#> [408] 2 1 1 3 2 2 1 1 1 1 1 3 1 2 3 1 2 1 1 2 1 3 1 3 3 3 1 2 2 1 1 1 3 3 2 3 1
#> [445] 3 2 1 1 1 1 3 1 3 2 1 3 2 2 2 2 1 3 3 3 2 2 2 3 3 2 1 1 1 3 3 1 1 3 1 2 3
#> [482] 2 2 1 3 2 3 2 2 1 1 1 1 1 3 3 1 1 3 1

## Get terminal nodes.
predictions <- predict(forests, X_test, type = "terminalNodes")
predictions$forest.1[1:10, 1:20] # Rows are observations, columns are forests.# }
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   29   41   61   27   15   37   25   80   29    29    27    57    35
#>  [2,]   33   24  106   74   19   38   76   68   92    90    30    67    16
#>  [3,]   92   20   93   88   36   23   94   64   95    98    88    12    62
#>  [4,]   24   40   39   41   52   49   22   18   14    40    14    12    33
#>  [5,]  108   40   39   41   23   47   36   93   61    23    75    12    78
#>  [6,]   28   30   43   43   34   49   22   18   14    40    54    63    13
#>  [7,]   62   13   57   90   82   61   84   63   88    96   100    12    78
#>  [8,]   62   13   57   90   82   49   84   78   88    96   100    12    78
#>  [9,]   81   62   41   93   86   83  106  100   67    47    34    62    84
#> [10,]   28   30   43   43   34   47   36   30   61    23    54    63    13
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]    51    27    29    15    14    15    85
#>  [2,]    53    48    51    23    75    64    71
#>  [3,]    46    65    26    72    44    46    82
#>  [4,]    46    24    72    56     6    75    22
#>  [5,]    45    63    82    69    45    75    55
#>  [6,]    49    24    44    56    45    39    11
#>  [7,]    46    66    70    72    77    76    55
#>  [8,]    46    66    70    72    77    76    55
#>  [9,]    75    42    32    61    16    79    83
#> [10,]    49    64    82    69    45    39    11