Tree Information in Readable Format

Extracts tree information from a morf.forest object.

tree_info(object, tree = 1)

Arguments

object

morf.forest object.

tree

Number of the tree of interest.

Value

A data.frame with the following columns:

nodeID

Node IDs.

leftChild

IDs of the left child node.

rightChild

IDs of the right child node.

splitvarID

IDs of the splitting variable.

splitvarName

Name of the splitting variable.

splitval

Splitting value.

terminal

Logical, TRUE for terminal nodes.

prediction

One column with the predicted conditional class probabilities.

Details

Nodes and variables IDs are 0-indexed, i.e., node 0 is the root node.

All values smaller than or equal to splitval go to the left and all values larger go to the right.

See also

Author

Riccardo Di Francesco

Examples

# \donttest{
## Load data from orf package.
set.seed(1986)

library(orf)
data(odata)

y <- as.numeric(odata[, 1])
X <- as.matrix(odata[, -1])

## Fit morf on training sample.
forests <- morf(y, X)

## Extract information from tenth tree of first forest.
info <- tree_info(forests$forests.info$forest.1, tree = 10)
info# }
#>     nodeID leftChild rightChild splitvarID     splitval terminal prediction
#> 1        0         1          2          0   0.13419280    FALSE         NA
#> 2        1         3          4          1   0.50000000    FALSE         NA
#> 3        2         5          6          0   1.01835785    FALSE         NA
#> 4        3         7          8          2   0.50000000    FALSE         NA
#> 5        4         9         10          2   0.50000000    FALSE         NA
#> 6        5        11         12          1   0.50000000    FALSE         NA
#> 7        6        13         14          3  16.08886646    FALSE         NA
#> 8        7        15         16          3  16.84751918    FALSE         NA
#> 9        8        NA         NA         NA           NA     TRUE  0.5588235
#> 10       9        17         18          1   1.50000000    FALSE         NA
#> 11      10        19         20          0  -1.47883449    FALSE         NA
#> 12      11        21         22          2   0.50000000    FALSE         NA
#> 13      12        23         24          0   0.20683760    FALSE         NA
#> 14      13        25         26          3  -0.58092994    FALSE         NA
#> 15      14        NA         NA         NA           NA     TRUE  0.2000000
#> 16      15        NA         NA         NA           NA     TRUE  0.9333333
#> 17      16        NA         NA         NA           NA     TRUE  0.0000000
#> 18      17        27         28          0  -1.54410947    FALSE         NA
#> 19      18        NA         NA         NA           NA     TRUE  0.2916667
#> 20      19        29         30          3   2.24716375    FALSE         NA
#> 21      20        31         32          3  -1.73132498    FALSE         NA
#> 22      21        33         34          0   0.20506876    FALSE         NA
#> 23      22        35         36          0   0.89861922    FALSE         NA
#> 24      23        37         38          0   0.16071925    FALSE         NA
#> 25      24        39         40          0   0.99589837    FALSE         NA
#> 26      25        NA         NA         NA           NA     TRUE  0.0000000
#> 27      26        41         42          0   1.54009584    FALSE         NA
#> 28      27        NA         NA         NA           NA     TRUE  1.0000000
#> 29      28        NA         NA         NA           NA     TRUE  0.5409836
#> 30      29        NA         NA         NA           NA     TRUE  1.0000000
#> 31      30        NA         NA         NA           NA     TRUE  0.7500000
#> 32      31        43         44          0  -0.58732707    FALSE         NA
#> 33      32        45         46          3   8.38977421    FALSE         NA
#> 34      33        NA         NA         NA           NA     TRUE  0.0000000
#> 35      34        47         48          0   0.40291238    FALSE         NA
#> 36      35        49         50          0   0.32614660    FALSE         NA
#> 37      36        NA         NA         NA           NA     TRUE  0.6666667
#> 38      37        NA         NA         NA           NA     TRUE  0.0000000
#> 39      38        51         52          2   0.50000000    FALSE         NA
#> 40      39        53         54          0   0.26235936    FALSE         NA
#> 41      40        NA         NA         NA           NA     TRUE  0.5000000
#> 42      41        55         56          2   0.50000000    FALSE         NA
#> 43      42        NA         NA         NA           NA     TRUE  0.0000000
#> 44      43        57         58          0  -0.77551070    FALSE         NA
#> 45      44        59         60          0  -0.09456219    FALSE         NA
#> 46      45        61         62          1   1.50000000    FALSE         NA
#> 47      46        63         64          1   1.50000000    FALSE         NA
#> 48      47        NA         NA         NA           NA     TRUE  1.0000000
#> 49      48        65         66          0   0.48951898    FALSE         NA
#> 50      49        NA         NA         NA           NA     TRUE  0.0000000
#> 51      50        67         68          3  -7.47771741    FALSE         NA
#> 52      51        NA         NA         NA           NA     TRUE  1.0000000
#> 53      52        NA         NA         NA           NA     TRUE  0.3333333
#> 54      53        69         70          0   0.25356577    FALSE         NA
#> 55      54        71         72          1   1.50000000    FALSE         NA
#> 56      55        73         74          0   1.33807743    FALSE         NA
#> 57      56        NA         NA         NA           NA     TRUE  0.0000000
#> 58      57        75         76          3 -10.67941890    FALSE         NA
#> 59      58        NA         NA         NA           NA     TRUE  0.7500000
#> 60      59        NA         NA         NA           NA     TRUE  0.0000000
#> 61      60        77         78          0  -0.09328891    FALSE         NA
#> 62      61        79         80          3   1.21926213    FALSE         NA
#> 63      62        81         82          3   0.34532435    FALSE         NA
#> 64      63        NA         NA         NA           NA     TRUE  0.2500000
#> 65      64        NA         NA         NA           NA     TRUE  0.0000000
#> 66      65        NA         NA         NA           NA     TRUE  0.0000000
#> 67      66        83         84          0   0.53809167    FALSE         NA
#> 68      67        NA         NA         NA           NA     TRUE  0.5000000
#> 69      68        85         86          0   0.42268685    FALSE         NA
#> 70      69        87         88          0   0.21858601    FALSE         NA
#> 71      70        NA         NA         NA           NA     TRUE  1.0000000
#> 72      71        89         90          0   0.35084562    FALSE         NA
#> 73      72        NA         NA         NA           NA     TRUE  0.0000000
#> 74      73        NA         NA         NA           NA     TRUE  0.0000000
#> 75      74        NA         NA         NA           NA     TRUE  0.6666667
#> 76      75        91         92          1   1.50000000    FALSE         NA
#> 77      76        NA         NA         NA           NA     TRUE  0.0000000
#> 78      77        NA         NA         NA           NA     TRUE  1.0000000
#> 79      78        NA         NA         NA           NA     TRUE  0.0000000
#> 80      79        93         94          3  -1.49274247    FALSE         NA
#> 81      80        95         96          3   3.72398033    FALSE         NA
#> 82      81        NA         NA         NA           NA     TRUE  0.0000000
#> 83      82        NA         NA         NA           NA     TRUE  0.7500000
#> 84      83        NA         NA         NA           NA     TRUE  1.0000000
#> 85      84        NA         NA         NA           NA     TRUE  0.6000000
#> 86      85        NA         NA         NA           NA     TRUE  1.0000000
#> 87      86        NA         NA         NA           NA     TRUE  0.0000000
#> 88      87        NA         NA         NA           NA     TRUE  0.0000000
#> 89      88        97         98          0   0.22895438    FALSE         NA
#> 90      89        NA         NA         NA           NA     TRUE  0.0000000
#> 91      90        99        100          0   0.35985115    FALSE         NA
#> 92      91        NA         NA         NA           NA     TRUE  0.3333333
#> 93      92        NA         NA         NA           NA     TRUE  0.0000000
#> 94      93        NA         NA         NA           NA     TRUE  1.0000000
#> 95      94       101        102          0  -0.32811118    FALSE         NA
#> 96      95        NA         NA         NA           NA     TRUE  0.0000000
#> 97      96        NA         NA         NA           NA     TRUE  0.3750000
#> 98      97        NA         NA         NA           NA     TRUE  1.0000000
#> 99      98        NA         NA         NA           NA     TRUE  0.0000000
#> 100     99        NA         NA         NA           NA     TRUE  1.0000000
#> 101    100       103        104          2   0.50000000    FALSE         NA
#> 102    101        NA         NA         NA           NA     TRUE  0.4000000
#> 103    102        NA         NA         NA           NA     TRUE  1.0000000
#> 104    103       105        106          0   0.84217485    FALSE         NA
#> 105    104       107        108          3   3.55888249    FALSE         NA
#> 106    105        NA         NA         NA           NA     TRUE  0.0000000
#> 107    106        NA         NA         NA           NA     TRUE  0.2500000
#> 108    107        NA         NA         NA           NA     TRUE  0.1764706
#> 109    108        NA         NA         NA           NA     TRUE  0.0000000